Handle landscape images.
Two changes were needing to support landscape images: The first was to remove the hardcoded check to ensure the dimensions, and the second was to run an object detector on the first frame to ensure that we can handle different size images without relying on the image dimensions. The object detection takes around 50ms to run, so running it on every frame to recover from a missing rounding box is prohibitive from an FPS perspective (though choosing a smaller model may improve latency). There's a chance the object detector isn't even needed on the first frame, but seems good hygiene to do it.

I'm choosing to embded the object model directly in here, since it's small and it simplifies deployment. It shouldn't be too much of a hardship on integrators. If it is, could look at a smaller model (e.g. lite0) or download it at runtime like the pose estimation model.